+++ src/lib.ts
@@ -78,8 +78,7 @@ export {
     MappedList,
     AsyncMappedList,
     ConcatList,
-    ObservableMap
-} from "./observable/index";
+} from "./observable";
 export {
     BaseObservableValue,
     ObservableValue,
